requirements.txt
asyncio==3.4.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
discord.py==2.1.0
dnspython==2.2.0
Flask==2.0.2
genshinstats==1.4.11.3
h11==0.12.0
idna==3.3
importlib-metadata==4.11.1
discord-py-interactions==4.3.4
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.5.0
lz4==4.0.0
MarkupSafe==2.0.1
multidict==5.2.0
outcome==1.1.0
pbkdf2==1.3
pyaes==1.6.1
pycparser==2.21
pycryptodome==3.14.1
pymongo==4.0.1
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pytz==2021.3
python-ffmpeg==1.0.14
pywin32-ctypes==0.2.0
requests==2.26.0
speedtest-cli==2.1.3
SecretStorage==3.3.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.7
user-agent==0.1.10
Werkzeug==2.0.2
wsproto==1.0.0
yarl==1.7.0
youtube-dl==2021.12.17
zipp==3.7.0
